feat(MOD-63): fixes at code by player mod files
